- Upgrade laravel 10 -> 11 - change form handling to spatie/laravel-html - add telescope

+++ app/Client.php
@@ -99,7 +99,7 @@ class Client extends Model
     public function Season()
     {
         $season = $this->seasonStart;
-        $season->year(Carbon::now()->format('Y'));
+        $season->year((int)Carbon::now()->format('Y'));
 
         if ($season->isFuture())
         {
